Change `:float` to `{:f, 64}` and add `{:f, 32}` dtype (#739)
* Change `:float` to `{:f, 64}` and add `{:f, 32}` dtype

The idea is to support more dtypes in the way they are represented
in Arrow/Polars.

In the near future, we won't need to "normalise" numeric columns anymore,
and by consequence, Explorer is going to use less memory if needed.

This is related to https://github.com/elixir-explorer/explorer/issues/471

* Add support for fetching from f32 series

* Make `Series.to_iovec/1` work for f32 series

* Make `Series.from_binary/3` work with "f32 binaries"

* Fix problems after rebase

* Fix mentions of `:float`

* Normalise dtypes instead of just validating

This is going to make easier to adopt "aliases" for dtypes.

* Make inspection of float series similar to Nx

Use the `"f64"` string instead of `float[64]`.

* Doc that `:float` can be used as alias for `{:f, 64}`

+++ test/explorer/backend/lazy_frame_test.exs
@@ -11,7 +11,7 @@ defmodule Explorer.Backend.LazyFrameTest do
              #Explorer.DataFrame<
                LazyFrame[??? x 2]
                a integer
-               b float
+               b f64
              >\
              """
   end
